2022_05_30_TowerDefence3+a
Fixed error
1. 건설 오류
2. 미사일 발사시 에러
3. 타워의 터렛이 적을 향하지 않는 오류
4. 타워 건설시 LevelManager의 돈이 증가하지 않는 오류
+5. 돈부족시 건설 못하는 멘트 찍기
+6. 건설 취소 버튼

+++ TowerDefence/Assets/02.Scripts/Tower/Tower_RocketLauncher.cs
@@ -5,10 +5,11 @@ using UnityEngine;
 public class Tower_RocketLauncher : Tower
 {
     public GameObject rocketPrefab;
-    public float damage;
+    public int damage;
     public float reloadTime;
     public float reloadTimer;
     [SerializeField] private Transform[] firePoints;
+
     protected override void Update()
     {
         base.Update();
